requirements.txt
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
pytest==3.3.2
text-unidecode==1.1
requests==2.18.4
python-dotenv==0.7.1
